+++ .idea/assetWizardSettings.xml
@@ -23,7 +23,7 @@
                                 <PersistentState>
                                   <option name="values">
                                     <map>
-                                      <entry key="url" value="file:/$USER_HOME$/AppData/Local/Android/Sdk/icons/material/materialicons/menu/baseline_menu_24.xml" />
+                                      <entry key="url" value="file:/$USER_HOME$/AppData/Local/Android/Sdk/icons/material/materialicons/bookmark/baseline_bookmark_24.xml" />
                                     </map>
                                   </option>
                                 </PersistentState>
@@ -34,7 +34,7 @@
                         <option name="values">
                           <map>
                             <entry key="color" value="00bcd4" />
-                            <entry key="outputName" value="ic_baseline_menu_24" />
+                            <entry key="outputName" value="ic_bookmark_48" />
                             <entry key="sourceFile" value="C:\Users\josar\AndroidStudioProjects\Foraneo" />
                           </map>
                         </option>
