+++ app/src/main/AndroidManifest.xml
@@ -20,6 +20,13 @@
                 android:value="" />
         </activity>
         <activity
+            android:name=".Search"
+            android:exported="true"> <!-- Aqui decia false -->
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+        </activity>
+        <activity
             android:name=".Login"
             android:exported="true"> <!-- Aqui decia false -->
