+++ app/src/main/res/drawable/ic_arrow_back_ios_new.xml
@@ -1,0 +1,5 @@
+<vector android:height="36dp" android:tint="#00BCD4"
+    android:viewportHeight="24" android:viewportWidth="24"
+    android:width="36dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="@android:color/white" android:pathData="M17.77,3.77l-1.77,-1.77l-10,10l10,10l1.77,-1.77l-8.23,-8.23z"/>
+</vector>
